+++ client/BalanceApp/components/user-list/user-list-item/user-list-item.js
@@ -1,56 +1,21 @@
 // vendors
 import React, { Component, PropTypes } from 'react';
-import { View, Text, Image, TouchableOpacity } from 'react-native';
-import { connect } from 'react-redux';
+import { View, Text, Image } from 'react-native';
 
+// styles
 import Styles from './user-list-item-styles';
 
-function mapStateToProps (state, ownProps) {
-  let fullLoggedInUser = state.users[state.loggedInUser];
+// components
+import FriendButton from '../../friend-button/friend-button';
 
-  let fromFriendList = fullLoggedInUser.friends.find(friend => {
-    return friend.userId === ownProps.user.userId;
-  });
-
-  return {
-    loggedInUser: state.loggedInUser,
-    friendStatus: fromFriendList && fromFriendList.status,
-    isFriend: fromFriendList ? true : false
-  }
-}
-
-class UserListItem extends Component {
+export default class UserListItem extends Component {
 
   static propTypes = {
     user: PropTypes.object.isRequired,
-    isFriend: PropTypes.bool.isRequired,
-    friendStatus: PropTypes.string,
-    loggedInUser: PropTypes.string.isRequired
   };
 
   constructor (props) {
     super(props);
-  }
-
-  friendActionText () {
-    const { friendStatus, user, loggedInUser, isFriend } = this.props;
-
-    if (user.userId === loggedInUser) {
-      return 'You!';
-    }
-
-    if (isFriend) {
-      switch (friendStatus) {
-        case 'requested':
-          return 'Accept';
-        case 'pending':
-          return 'Cancel Request';
-        default:
-          return 'Remove';
-      }
-    }
-    
-    return 'Add';
   }
 
   renderUsername () {
@@ -67,7 +32,7 @@ class UserListItem extends Component {
   }
 
   render () {
-    const { user, loggedInUser, isFriend } = this.props;
+    const { user } = this.props;
     return (
       <View style={Styles.userListItem} >
         <Image source={{ uri: user.picture }} style={Styles.picture} />
@@ -75,13 +40,9 @@ class UserListItem extends Component {
           <Text style={Styles.text} >{user.name}</Text>
           { this.renderUsername() }
         </View>
-        <TouchableOpacity>
-          <Text style={Styles.lighter}>{ this.friendActionText() }</Text>
-        </TouchableOpacity>
+        <FriendButton userId={ user.userId } />
       </View>
     );
   }
 
 }
-
-export default connect(mapStateToProps)(UserListItem);
